# Conflicts:
#	.claude/rules/core.md
#	CLAUDE.md
#	README.md
#	docs/ai/change-trace.md
#	docs/ai/decision-log.md
#	docs/ai/mistakes.md
#	docs/design/review-2026-07-15.md
#	include/api/static_assets.h
#	include/cli/banner.h
#	src/api/static_assets.c
#	src/cli/banner.c
#	src/cli/main.c
#	tests/test_static_assets.c
#	tools/screenshots/cli/capture.html
#	tools/screenshots/cli/capture.mjs

+++ tests/test_vision_e2e.c
@@ -37,7 +37,7 @@ static int64_t now_us(void) {
 /* ── tiny random weight init ─────────────────────────────────────────────── */
 static void init_weights_random(float *w, int n) {
     for (int i = 0; i < n; i++)
-        w[i] = ((float)rand() / RAND_MAX - 0.5f) * 0.02f;
+        w[i] = ((float)rand() / (float)RAND_MAX - 0.5f) * 0.02f;
 }
 static void init_weights_ones(float *w, int n) {
     for (int i = 0; i < n; i++) w[i] = 1.0f;
